Added libssl1.1 fallback installation

+++ SetupMinecraft.sh
@@ -157,15 +157,19 @@ Check_Dependencies() {
     if [[ "$SSLVer" ]]; then
       sudo DEBIAN_FRONTEND=noninteractive apt-get install libssl1.1 -yqq
     else
-      echo "No libssl1.1 available in repositories -- attempting manual install"
-      sudo curl -o libssl.deb -k -L http://archive.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1l-1ubuntu1.2_amd64.deb
-      sudo dpkg -i libssl.deb
-      sudo rm libssl.deb
-      SSLVer=$(apt-cache show libssl1.1 | grep Version | awk 'NR==1{ print $2 }')
-      if [[ "$SSLVer" ]]; then
-        echo "Manual libssl1.1 installation successful!"
-      else
-        echo "Manual libssl1.1 installation failed."
+      CPUArch=$(uname -m)
+      if [[ "$CPUArch" == *"x86_64"* ]]; then
+        echo "No libssl1.1 available in repositories -- attempting manual install"
+        
+        sudo curl -o libssl.deb -k -L http://archive.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1l-1ubuntu1.2_amd64.deb
+        sudo dpkg -i libssl.deb
+        sudo rm libssl.deb
+        SSLVer=$(apt-cache show libssl1.1 | grep Version | awk 'NR==1{ print $2 }')
+        if [[ "$SSLVer" ]]; then
+          echo "Manual libssl1.1 installation successful!"
+        else
+          echo "Manual libssl1.1 installation failed."
+        fi
       fi
     fi
